test: BetaRndの乱数テストをクラスに分離
