+++ test/matsu/num/statistics/random/beta/GammaBasedBetaRndTest.java
@@ -10,18 +10,13 @@ import static matsu.num.statistics.random.BetaRnd.*;
 import static org.hamcrest.MatcherAssert.*;
 import static org.hamcrest.Matchers.*;
 
-import java.util.Objects;
-
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.Test.None;
 import org.junit.experimental.runners.Enclosed;
 import org.junit.runner.RunWith;
 
-import matsu.num.statistics.random.BaseRandom;
 import matsu.num.statistics.random.BetaRnd;
 import matsu.num.statistics.random.FloatingRandomGeneratorTestingFramework;
-import matsu.num.statistics.random.TestedFloatingRandomGenerator;
 import matsu.num.statistics.random.gamma.GammaFactoryForTesting;
 
 /**
@@ -83,88 +78,21 @@ final class GammaBasedBetaRndTest {
 
     public static class ベータ乱数のテスト {
 
-        private FloatingRandomGeneratorTestingFramework framework;
-
-        @Before
-        public void before() {
-            framework =
-                    FloatingRandomGeneratorTestingFramework
-                            .instanceOf(new TestedBeta(BaseRandom.threadSeparatedRandom()));
-        }
-
         @Test
         public void test() {
-            framework.test();
-        }
-
-        /**
-         * a=2,b=1のベータ乱数発生器のテスタ. <br>
-         * 累積分布関数はx^2
-         */
-        private static final class TestedBeta implements TestedFloatingRandomGenerator {
-
-            private final BaseRandom random;
-            private final BetaRnd betaRnd;
-
-            public TestedBeta(BaseRandom random) {
-                this.random = Objects.requireNonNull(random);
-                this.betaRnd = FACTORY.instanceOf(2, 1);
-            }
-
-            @Override
-            public double newValue() {
-                return betaRnd.nextRandom(random);
-            }
-
-            @Override
-            public double cumulativeProbability(double arg) {
-                return arg * arg;
-            }
-
+            FloatingRandomGeneratorTestingFramework
+                    .instanceOf(new Tested_At_2_1_BetaRandomGenerator(FACTORY))
+                    .test();
         }
     }
 
     public static class ベータプライム乱数のテスト {
 
-        private FloatingRandomGeneratorTestingFramework framework;
-
-        @Before
-        public void before() {
-            framework =
-                    FloatingRandomGeneratorTestingFramework
-                            .instanceOf(new TestedBetaPrime(BaseRandom.threadSeparatedRandom()));
-        }
-
         @Test
         public void test() {
-            framework.test();
-        }
-
-        /**
-         * a=2,b=1のベータプライム乱数発生器のテスタ. <br>
-         * 累積分布関数は<br>
-         * {@literal 1 - (1+2x)/(1+x)^2}
-         */
-        private static final class TestedBetaPrime implements TestedFloatingRandomGenerator {
-
-            private final BaseRandom random;
-            private final BetaRnd betaRnd;
-
-            public TestedBetaPrime(BaseRandom random) {
-                this.random = Objects.requireNonNull(random);
-                this.betaRnd = FACTORY.instanceOf(2, 1);
-            }
-
-            @Override
-            public double newValue() {
-                return betaRnd.nextBetaPrime(random);
-            }
-
-            @Override
-            public double cumulativeProbability(double arg) {
-                return 1 - (1 + 2 * arg) / ((1 + arg) * (1 + arg));
-            }
-
+            FloatingRandomGeneratorTestingFramework
+                    .instanceOf(new Tested_At_2_1_BetaPrimeRandomGenerator(FACTORY))
+                    .test();
         }
     }
